+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* [improvement] git remote created after `shelly add` and `shelly setup` is named 'shelly'
+
 ## 0.2.21 / 2013-05-13
 
 * [improvement] require 'puma' or 'thin' gems if present in Cloudfile when running `shelly check`